fix(core): add readonly to lifecycle field, extract shared createRunbook fixture
- Mark RunbookContext.lifecycle as readonly to match interface convention
- Extract createRunbook helper to shared fixtures.ts
- Update lifecycle-context.test.ts and compiler.test.ts to import from fixtures
